+++ app/routes/redirect.py
@@ -13,11 +13,10 @@ from itsdangerous import BadSignature
 from werkzeug import Response
 
 from app.forms import EmptyForm
-from app.models import Post, User, db
+from app.models import User, db
 from app.utils import redirect
 from app.utils.mail import send_email
 from app.utils.security import (
-    authorization_required,
     confirm_token,
     confirmation_required,
     generate_confirmation_token,
@@ -79,25 +78,6 @@ def resend_confirmation() -> Response:
     )
     flash(_("A new confirmation email has been sent."))
     return redirect.Auth.unconfirmed()
-
-
-@blueprint.route("/<int:id>/delete", methods=["POST"])
-@login_required
-@authorization_required
-def delete(id: int) -> Response:
-    """Delete post by post's ID.
-
-    The "delete" view does not have its own template. Since there is no
-    template it will only handle the ``POST`` method and then redirect
-    to the index view.
-
-    :param id: The post's ID.
-    :return: Response object redirect to index view.
-    """
-    post = Post.get_post(id)
-    db.session.delete(post)
-    db.session.commit()
-    return redirect.index()
 
 
 @blueprint.route("/follow/<username>", methods=["POST"])