+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/jss/compare/v1.19.1...HEAD)
 ------------------------------------------------------------------------
+### Changed
+- Moves `/redirect/<int:id>/delete` to `/post/<int:id>/delete`
 
 [1.19.1](https://github.com/jshwi/jss/releases/tag/v1.19.1) - 2022-06-04
 ------------------------------------------------------------------------